Decode Ways
This problem is usually solved by using dynamic programming.

In this solution, I used a different method to solve this problem other
than dynamic programming.
Firstly, divide the string into sub segments,
and then multiply the ways of each segments.
Use an array to store the possible ways of different length, which are
Fibonacci numbers.